ISSUE-1086: Ability to view large messages and limit the message size for log search output
