+++ webservice/src/main/resources/app/scripts/containers/LogSearch/LogSearch.jsx
@@ -36,12 +36,14 @@ import {
 import FSReactToastr from '../../components/FSReactToastr';
 import BaseContainer from '../../containers/BaseContainer';
 import CommonNotification from '../../utils/CommonNotification';
+import ParagraphShowHideComponent from '../../utils/ParagraphShowHide';
 import {toastOpt} from '../../utils/Constants';
 import LogSearchREST from '../../rest/LogSearchREST';
 import moment from 'moment';
-import DatetimeRangePicker from 'react-bootstrap-datetimerangepicker';
 import {Select2 as Select} from '../../utils/SelectUtils';
 import TopologyREST from '../../rest/TopologyREST';
+import DateTimePickerDropdown from '../../components/DateTimePickerDropdown';
+import TablePagination from '../../components/TablePagination';
 
 class LogSearch extends Component {
   constructor(props) {
@@ -73,40 +75,6 @@ class LogSearch extends Component {
       noOfResults: 0,
       startDate: moment().subtract(30, 'minutes'),
       endDate: moment(),
-      ranges: {
-        'Last 30 Minutes': [
-          moment().subtract(30, 'minutes'),
-          moment()
-        ],
-        'Last 1 Hour': [
-          moment().subtract(1, 'hours'),
-          moment()
-        ],
-        'Last 3 Hours': [
-          moment().subtract(3, 'hours'),
-          moment()
-        ],
-        'Last 6 Hours': [
-          moment().subtract(6, 'hours'),
-          moment()
-        ],
-        'Last 12 Hours': [
-          moment().subtract(12, 'hours'),
-          moment()
-        ],
-        'Last 24 Hours': [
-          moment().subtract(1, 'days'),
-          moment()
-        ],
-        'Last 7 Days': [
-          moment().subtract(6, 'days'),
-          moment()
-        ],
-        'Last 30 Days': [
-          moment().subtract(29, 'days'),
-          moment()
-        ]
-      },
       showColumn:{
         timestamp: true,
         logLevel: true,
@@ -119,7 +87,6 @@ class LogSearch extends Component {
   }
   componentDidMount(){
     this.fetchComponents();
-    this.fetchLogs();
   }
   getHeaderContent() {
     const {topologyData} = this.state;
@@ -137,7 +104,11 @@ class LogSearch extends Component {
       return '';
     }
   }
+  getSelectedViewModeComponent = (componentId,nodes) => {
+    return _.filter(nodes, (node) => node.id === componentId);
+  }
   fetchComponents(){
+    const {componentId=''} =  this.props.location.state ? this.props.location.state : {};
     const topologyId = this.props.routeParams.id;
 
     TopologyREST.getTopology(topologyId, undefined).then((data) => {
@@ -149,13 +120,18 @@ class LogSearch extends Component {
       promiseArr.push(TopologyREST.getAllNodes(topologyId, versionId, 'sinks'));
 
       Promise.all(promiseArr).then((resultsArr) => {
-        const nodes = [];
+        const nodes = [],stateObj={};
         nodes.push.apply(nodes, resultsArr[0].entities);
         nodes.push.apply(nodes, resultsArr[1].entities);
         nodes.push.apply(nodes, resultsArr[2].entities);
-        this.setState({
-          topologyData: data,
-          components: nodes
+        let tempSelectedComp=[];
+        if(componentId !== ''){
+          stateObj.selectedComponents = this.getSelectedViewModeComponent(componentId,nodes);
+        }
+        stateObj.topologyData = data;
+        stateObj.components = nodes;
+        this.setState(stateObj, () => {
+          this.fetchLogs();
         });
       });
     });
@@ -207,17 +183,17 @@ class LogSearch extends Component {
       });
     });
   }
-  handlePagination = (eventKey) => {
+  paginationCallback = (eventKey) => {
     this.setState({
       activePage: eventKey
     }, () => {
       this.fetchLogs();
     });
   }
-  handleEvent = (e, datePicker) => {
+  datePickerCallback = (startDate, endDate) => {
     this.setState({
-      startDate: datePicker.startDate,
-      endDate: datePicker.endDate,
+      startDate: startDate,
+      endDate: endDate,
       activePage: 1
     }, () => {
       this.fetchLogs();
@@ -252,11 +228,6 @@ class LogSearch extends Component {
   onSearchBtnClick = () => {
     this.fetchLogs();
   }
-  getActivePageData = (allData, activePage, pageSize) => {
-    activePage = activePage - 1;
-    const startIndex = activePage*pageSize;
-    return allData.slice(startIndex, startIndex+pageSize);
-  }
   getLogLabel(logLevel){
     switch(logLevel){
     case 'DEBUG':
@@ -278,38 +249,74 @@ class LogSearch extends Component {
   }
   getDateTime(timestamp){
     return <div>
-      <span>{moment(timestamp).format('MM-DD-YYYY')}</span><br />
-      <span>{moment(timestamp).format('HH:mm')}</span>
+      <span title={moment(timestamp).format('MM-DD-YYYY') + '  ' + moment(timestamp).format('HH:mm')}>
+        {moment(timestamp).fromNow()}
+      </span>
     </div>;
   }
   getTableHeaderContent(){
     const {showColumn} = this.state;
     const content = [];
-    showColumn.timestamp ? content.push(<Th key="timestamp" column="timestamp" style={{minWidth: '90px'}}>DateTime</Th>) : null;
+    const columnSelectButton = (
+      <DropdownButton
+        id="column-select"
+        title={<i className="fa fa-cog" aria-hidden="true"></i>}
+        pullRight={true}
+        onSelect={this.onColumnSelect}
+        bsStyle="link"
+      >
+        <MenuItem eventKey="timestamp">
+          {showColumn.timestamp ?
+            <i className="fa fa-check-square-o" aria-hidden="true"></i>
+            : <i className="fa fa-square-o" aria-hidden="true"></i>} DateTime
+        </MenuItem>
+        <MenuItem eventKey="logLevel">
+          {showColumn.logLevel ?
+            <i className="fa fa-check-square-o" aria-hidden="true"></i>
+            : <i className="fa fa-square-o" aria-hidden="true"></i>} Log Level
+        </MenuItem>
+        <MenuItem eventKey="componentName">
+          {showColumn.componentName ?
+            <i className="fa fa-check-square-o" aria-hidden="true"></i>
+            : <i className="fa fa-square-o" aria-hidden="true"></i>} Component Name
+        </MenuItem>
+        <MenuItem eventKey="host">
+          {showColumn.host ?
+            <i className="fa fa-check-square-o" aria-hidden="true"></i>
+            : <i className="fa fa-square-o" aria-hidden="true"></i>} Host
+        </MenuItem>
+        <MenuItem eventKey="port">
+          {showColumn.port ?
+            <i className="fa fa-check-square-o" aria-hidden="true"></i>
+            : <i className="fa fa-square-o" aria-hidden="true"></i>} Port
+        </MenuItem>
+      </DropdownButton>
+    );
+    showColumn.timestamp ? content.push(<Th key="timestamp" column="timestamp" style={{minWidth: '90px'}}>Date<span>/</span>Time</Th>) : null;
     showColumn.logLevel ? content.push(<Th key="logLevel" column="logLevel">Log Level</Th>) : null;
     showColumn.componentName ? content.push(<Th key="componentName" column="componentName">Component Name</Th>) : null;
     showColumn.host ? content.push(<Th key="host" column="host">Host</Th>) : null;
     showColumn.port ? content.push(<Th key="port" column="port">Port</Th>) : null;
-    showColumn.logMessage ? content.push(<Th key="logMessage" column="logMessage">Log Message</Th>) : null;
+    showColumn.logMessage ? content.push(<Th key="logMessage" column="logMessage">Log Message<div className="pull-right">{columnSelectButton}</div></Th>) : null;
     return content;
   }
-  getRowContent(log){
+  getRowContent(log,i){
     const {showColumn} = this.state;
     const content = [];
-    showColumn.timestamp ? content.push(<Td column="timestamp">{this.getDateTime(log.timestamp)}</Td>) : null;
-    showColumn.logLevel ? content.push(<Td column="logLevel">{this.getLogLabel(log.logLevel)}</Td>) : null;
-    showColumn.componentName ? content.push(<Td column="componentName"><span className="text-info">{log.componentName}</span></Td>) : null;
-    showColumn.host ? content.push(<Td column="host">{log.host}</Td>) : null;
-    showColumn.port ? content.push(<Td column="port">{log.port}</Td>) : null;
-    showColumn.logMessage ? content.push(<Td column="logMessage">{log.logMessage}</Td>) : null;
+    showColumn.timestamp ? content.push(<Td key={'timestamp_'+i} column="timestamp">{this.getDateTime(log.timestamp)}</Td>) : null;
+    showColumn.logLevel ? content.push(<Td key={'logLevel_'+i} column="logLevel">{this.getLogLabel(log.logLevel)}</Td>) : null;
+    showColumn.componentName ? content.push(<Td key={'componentName_'+i} column="componentName"><span className="text-info">{log.componentName}</span></Td>) : null;
+    showColumn.host ? content.push(<Td key={'host_'+i} column="host">{log.host}</Td>) : null;
+    showColumn.port ? content.push(<Td key={'port_'+i} column="port">{log.port}</Td>) : null;
+    showColumn.logMessage ? content.push(<Td key={'logMessage_'+i} column="logMessage" className="wordBreak"><ParagraphShowHideComponent linkClass="log-message" showMaxText={300} content={log.logMessage} /></Td>) : null;
     return content;
   }
+
   render() {
     const {
       logs,
       activePage,
       pageSize,
-      ranges,
       startDate,
       endDate,
       logLevels,
@@ -350,101 +357,63 @@ class LogSearch extends Component {
         <div className="row">
           <div className="col-sm-12">
             <div style={{background: 'white', padding: '20px'}}>
-              <Panel header={<h3><i className="fa fa-filter" aria-hidden="true"></i> Filters</h3>}>
-                <div className="row">
-                  <FormGroup className="col-sm-6">
-                    <label>Component :</label>
-                    <Select
-                      name="log-level"
-                      value={selectedComponents}
-                      options={components}
-                      labelKey="name"
-                      valueKey="name"
-                      placeholder="Select Component"
-                      multi={true}
-                      onChange={this.handleComponentChange}
-                    />
-                  </FormGroup>
-                  <FormGroup className="col-sm-6">
-                    <label>Search :</label>
-                    <div className="input-group add-on">
-                      <input value={searchString} ref="searchInput" onChange={this.handleSearchStringChange} onKeyUp={this.handleSearchStringKeyUp} className="form-control" placeholder="Search" name="srch-term" id="srch-term" type="text" />
-                      <div className="input-group-btn">
-                        <button className="btn btn-default" onClick={this.onSearchBtnClick}><i className="fa fa-search"></i></button>
-                      </div>
-                    </div>
-                  </FormGroup>
+              <div className="row">
+                <FormGroup className="col-sm-6">
+                  <label>Component</label>
+                  <Select
+                    name="log-level"
+                    value={selectedComponents}
+                    options={components}
+                    labelKey="name"
+                    valueKey="name"
+                    placeholder="Select Component"
+                    multi={true}
+                    onChange={this.handleComponentChange}
+                  />
+                </FormGroup>
+                <FormGroup className="col-sm-6">
+                  <label>Log Level</label>
+                  <Select
+                    name="log-level"
+                    value={logLevels}
+                    options={logLevelOptions}
+                    placeholder="Select Log Level"
+                    multi={true}
+                    onChange={this.handleLogLevelChange}
+                  />
+                </FormGroup>
+              </div>
+              <div className="clearfix form-group nomargin">
+                <label>Search</label>
+              </div>
+              <div className="row">
+                <div className="col-sm-12 form-group">
+                  <div className="input-group">
+                    <input value={searchString} ref="searchInput" onChange={this.handleSearchStringChange} onKeyUp={this.handleSearchStringKeyUp} className="form-control" placeholder="Search" name="search-term" id="search-term" type="text" />
+                    <span className="input-group-btn">
+                      <DateTimePickerDropdown
+                      dropdownId="log-search-datepicker-dropdown"
+                      startDate={startDate}
+                      endDate={endDate}
+                      locale={locale}
+                      isDisabled={false}
+                      datePickerCallback={this.datePickerCallback} />
+                      <Button className="btn btn-success log-search-btn" onClick={this.onSearchBtnClick}><i className="fa fa-search"></i></Button>
+                    </span>
+                  </div>
                 </div>
-                <div className="row">
-                  <FormGroup className="col-sm-6">
-                    <label>Log Level :</label>
-                    <Select
-                      name="log-level"
-                      value={logLevels}
-                      options={logLevelOptions}
-                      placeholder="Select Log Level"
-                      multi={true}
-                      onChange={this.handleLogLevelChange}
-                    />
-                  </FormGroup>
-                  <FormGroup className="col-sm-6">
-                    <label>Range :</label>
-                    <DatetimeRangePicker timePicker timePicker24Hour showDropdowns timePickerSeconds locale={locale} startDate={startDate} endDate={endDate} ranges={ranges} onApply={this.handleEvent} opens="left">
-                      <div className="input-group">
-                        <input type="text" className="form-control" defaultValue={label}/>
-                        <span className="input-group-btn">
-                          <Button className="default date-range-toggle">
-                            <i className="fa fa-calendar"/>
-                          </Button>
-                        </span>
-                      </div>
-                    </DatetimeRangePicker>
-                  </FormGroup>
-                </div>
-              </Panel>
+                {/*<div className="col-sm-1 text-right">
+                  <Button className="btn btn-default" style={{padding: '4px', width: '35px', filter: 'grayscale(100%)'}}>
+                    <img src="/styles/img/icon-ambari_metrics.png" width="100%"/>
+                  </Button>
+                </div>*/}
+              </div>
               <div className="row text-right" style={{marginBottom: '5px'}}>
-                <DropdownButton
-                  id="column-select"
-                  title={<i className="fa fa-cog" aria-hidden="true"></i>}
-                  pullRight={true}
-                  onSelect={this.onColumnSelect}
-                >
-                  <MenuItem eventKey="timestamp">
-                    {showColumn.timestamp ?
-                      <i className="fa fa-check-square-o" aria-hidden="true"></i>
-                      : <i className="fa fa-square-o" aria-hidden="true"></i>} DateTime
-                  </MenuItem>
-                  <MenuItem eventKey="logLevel">
-                    {showColumn.logLevel ?
-                      <i className="fa fa-check-square-o" aria-hidden="true"></i>
-                      : <i className="fa fa-square-o" aria-hidden="true"></i>} Log Level
-                  </MenuItem>
-                  <MenuItem eventKey="componentName">
-                    {showColumn.componentName ?
-                      <i className="fa fa-check-square-o" aria-hidden="true"></i>
-                      : <i className="fa fa-square-o" aria-hidden="true"></i>} Component Name
-                  </MenuItem>
-                  <MenuItem eventKey="host">
-                    {showColumn.host ?
-                      <i className="fa fa-check-square-o" aria-hidden="true"></i>
-                      : <i className="fa fa-square-o" aria-hidden="true"></i>} Host
-                  </MenuItem>
-                  <MenuItem eventKey="port">
-                    {showColumn.port ?
-                      <i className="fa fa-check-square-o" aria-hidden="true"></i>
-                      : <i className="fa fa-square-o" aria-hidden="true"></i>} Port
-                  </MenuItem>
-                  {/*<MenuItem eventKey="logMessage">
-                    {showColumn.logMessage ?
-                      <i className="fa fa-check-square-o" aria-hidden="true"></i>
-                      : <i className="fa fa-square-o" aria-hidden="true"></i>} Log Message
-                  </MenuItem>*/}
-                </DropdownButton>
               </div>
               <div className="row">
                 <div className="table-responsive">
                   <Table
-                    className="table table-hover table-stream"
+                    className="table table-hover table-stream log-search-table"
                     noDataText="No records found."
                   >
                     <Thead>
@@ -459,18 +428,11 @@ class LogSearch extends Component {
                 </div>
               </div>
               <div className="row">
-                <Pagination
-                  className={`${noOfResults === 0? 'hidden':'shown'} pull-right`}
-                  prev={<i className="fa fa-chevron-left"></i>}
-                  next={<i className="fa fa-chevron-right"></i>}
-                  first={<i className="fa fa-angle-double-left"></i>}
-                  last={<i className="fa fa-angle-double-right"></i>}
-                  ellipsis
-                  items={Math.ceil(noOfResults/pageSize)}
-                  maxButtons={5}
+                <TablePagination
                   activePage={activePage}
-                  onSelect={this.handlePagination}>
-                </Pagination>
+                  pageSize={pageSize}
+                  noOfResults={noOfResults}
+                  paginationCallback={this.paginationCallback} />
               </div>
             </div>
           </div>
