+++ webservice/src/main/resources/app/scripts/utils/CommonNotification.jsx
@@ -13,29 +13,18 @@
 **/
 import React, {Component} from 'react';
 import {notifyTextLimit} from '../utils/Constants';
+import ParagraphShowHideComponent from '../utils/ParagraphShowHide';
 
-class CommonNotification extends Component {
-  constructor(props) {
+class CommonNotification extends ParagraphShowHideComponent {
+  constructor(props){
     super(props);
-    this.state = {
-      data: false,
-      text: "Read more"
-    };
-  }
-  showMore = () => {
-    if (this.state.text === "Read more") {
-      this.setState({text: "Hide", data: true});
-    } else {
-      this.setState({text: "Read more", data: false});
-    }
   }
 
   render() {
-    /* flag value         error, info, sucess */
+    // flag value         error, info, sucess
     const {text, data} = this.state;
     const {flag, content} = this.props;
-    const initial = content.substr(0, notifyTextLimit);
-    const moreText = content.substr(notifyTextLimit);
+    const {initial,moreText} = this.getIntialTextFromContent(content);
     const readMoreTag = <a href="javascript:void(0)" onClick={this.showMore}>{text}</a>;
     return (
       <div>
@@ -55,9 +44,9 @@ class CommonNotification extends Component {
   }
 }
 
-export default CommonNotification;
-
 CommonNotification.propTypes = {
   flag: React.PropTypes.string.isRequired,
   content: React.PropTypes.string
 };
+
+export default CommonNotification;
